[nrf noup] manifest: Manual revert of Upgrade to trusted-firmware-m 1.8.0
Currently sdk-nrf is not aligned with TF-M 1.8.

This commit manually reverts the:

commit 3b3e28aedaf8f30b03eb1e1753d9ab34d34c5d99
Author: David Brown <david.brown@linaro.org>
Date:   Mon Jul 10 12:12:21 2023 -0600

manifest: Upgrade to trusted-firmware-m 1.8.0

The manual revert, which is done by applying negative patch,
has been done because we are reverting commit that has been brought
as part of merge; if we would later like to have that commit it
would be hard to revert real revert commit, additionally git type
reverts may not get detected during next merge which will cause
history to get broken.

This commit can be easily reverted later, with no harm for history.

Signed-off-by: Dominik Ermel <dominik.ermel@nordicsemi.no>